Updated Boolean and Goedel logics; test + minor fixes (see decription)
Updated Boolean and Goedel logics to indexed version. Added warning to  constructor using  when the operation is not commutative. Added conversion to  in  function notation calls to allow use with . Fixed small bugs in  and  constructors (and relative axioms). Updated  documentation. Fixed small bugs in many-valued check. Updated tests. Commented old or working on lines.